backend/app/requirements.txt
    fastapi[standard]<1.0.0,>=0.115.6,
    pydantic>2.9.2,
    python-multipart<1.0.0,>=0.0.17,
    PyJWT<3.0.0,>=2.10.0,
    passlib[bcrypt]<2.0.0,>=1.7.4,
    bcrypt==4.2.1,
    sqlmodel==0.0.22,
    pydantic-settings<3.0.0,>=2.6.1,
  
  